+++ lib/mini_game/components/MiniGameTestBg.dart
