+++ lib/widgets/DistanceBar.dart
